+++ src/main/java/webproject/ticket/domain/member/Member.java
@@ -2,32 +2,23 @@ package webproject.ticket.domain.member;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
-import javax.validation.constraints.NotBlank;
-
 @Getter
 @Setter
 @Entity
 @AllArgsConstructor
-@NoArgsConstructor
 public class Member {
 
     @Id
     @GeneratedValue
     private Long id;
-
-    @Column(name = "member_id")
-    @NotBlank
-    private String memberId;
-
+    private String member_id;
     //private String name;
-    @NotBlank
     private String pw;
     //private String email;
