+++ src/main/java/webproject/ticket/domain/order/OrderRepository.java
@@ -1,4 +1,4 @@
-/*package webproject.ticket.domain.order;
+package webproject.ticket.domain.order;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
@@ -20,4 +20,3 @@ public class OrderRepository {
         return em.find(Order.class, id);
     }
 }
-*/
