+++ src/main/resources/templates/index.html
@@ -4,11 +4,10 @@
 <div class="container">
     <div>
         <h1>Hello Spring</h1>
-        <p>환영합니다, 손님.</p>
         <p>회원 기능</p>
         <p>
-            <a href="members/sign-up">회원가입</a>
-            <a href="members/sign-in">로그인</a>
+            <a href="member/signup">회원가입</a>
+            <a href="member/signin">로그인</a>
         </p>
     </div>
 </div> <!-- /container -->
