+++ src/main/resources/templates/member/signup.html
@@ -2,10 +2,10 @@
 <html xmlns:th="http://www.thymeleaf.org">
 <body>
 <div class="container">
-    <form action="/members/sign-up" method="post">
+    <form action="/member/signup" method="post">
         <div class="form-group">
-            <label for="memberId">ID</label>
-            <input type="text" id="memberId" name="memberId">
+            <label for="member_id">ID</label>
+            <input type="text" id="member_id" name="member_id">
             <label for="pw">PW</label>
             <input type="text" id="pw" name="pw">
         </div>
