--- src/main/resources/templates/members/sign-in-home.html
@@ -1,16 +1,0 @@
-<!DOCTYPE HTML>
-<html xmlns:th="http://www.thymeleaf.org">
-<body>
-<div class="container">
-    <div>
-        <h1>Hello Spring</h1>
-        <p th:text="'환영합니다. ' + ${memberId} + '님'"></p>
-        <p>회원 기능</p>
-        <p>
-            <a href="members/sign-up">회원가입</a>
-            <a href="members/sign-out">로그아웃</a>
-        </p>
-    </div>
-</div> <!-- /container -->
-</body>
-</html>
